spike: spring boot test for management controller
